refactor: name에 대한 검증 Name 도메인에 위임
