+++ src/main/java/wooteco/subway/domain/Station.java
@@ -4,57 +4,17 @@ import java.util.Objects;
 
 public class Station {
 
-    private static final String ERROR_NULL_OR_EMPTY = "[ERROR] 이름에 빈칸 입력은 허용하지 않습니다.";
-    private static final String BLANK = " ";
-    private static final String ERROR_INCLUDE_BLANK = "[ERROR] 이름은 내부에 공백이 포함될 수 없습니다.";
-    private static final String ERROR_INVALID_LENGTH = "[ERROR] 이름은 2글자 이상이어야합니다.";
-    private static final String REGEX_SPECIAL = "[가-힣\\w_]+";
-    private static final String ERROR_SPECIAL = "[ERROR] 이름에 특수문자를 입력할 수 없습니다.";
 
     private Long id;
-    private String name;
+    private Name name;
 
     public Station(final String name) {
         this(null, name);
     }
 
     public Station(final Long id, String name) {
-        validateName(name);
         this.id = id;
-        this.name = name;
-    }
-
-    private void validateName(final String name) {
-        checkNullOrEmpty(name);
-        checkIncludeBlank(name);
-        checkValidLengthOfName(name);
-        checkSpecialCharInName(name);
-    }
-
-    private void checkNullOrEmpty(final String name) {
-        Objects.requireNonNull(name, ERROR_NULL_OR_EMPTY);
-        if (name.trim().isEmpty()) {
-            throw new IllegalArgumentException(ERROR_NULL_OR_EMPTY);
-        }
-    }
-
-    private void checkIncludeBlank(final String name) {
-        if (name.trim().contains(BLANK)) {
-            throw new IllegalArgumentException(ERROR_INCLUDE_BLANK);
-        }
-    }
-
-    private void checkValidLengthOfName(final String name) {
-        int nameLength = name.length();
-        if (2 > nameLength) {
-            throw new IllegalArgumentException(ERROR_INVALID_LENGTH);
-        }
-    }
-
-    private void checkSpecialCharInName(final String name) {
-        if (!name.matches(REGEX_SPECIAL)) {
-            throw new IllegalArgumentException(ERROR_SPECIAL);
-        }
+        this.name = new Name(name);
     }
 
     public Long getId() {
@@ -62,7 +22,7 @@ public class Station {
     }
 
     public String getName() {
-        return name;
+        return name.getValue();
     }
 
     @Override
